+++ src/test/java/co/fineants/api/docs/member/MemberNotificationRestControllerDocsTest.java
@@ -20,16 +20,16 @@ import org.springframework.restdocs.payload.JsonFieldType;
 
 import co.fineants.TestDataFactory;
 import co.fineants.api.docs.RestDocsSupport;
-import co.fineants.api.domain.member.controller.MemberNotificationRestController;
-import co.fineants.member.application.MemberNotificationPreferenceService;
-import co.fineants.member.application.MemberNotificationService;
 import co.fineants.api.domain.notification.domain.entity.PortfolioNotification;
 import co.fineants.api.domain.notification.domain.entity.StockTargetPriceNotification;
 import co.fineants.api.domain.portfolio.domain.entity.Portfolio;
 import co.fineants.api.domain.stock.domain.entity.Stock;
 import co.fineants.api.domain.stock_target_price.domain.entity.TargetPriceNotification;
 import co.fineants.api.global.util.ObjectMapperUtil;
+import co.fineants.member.application.MemberNotificationPreferenceService;
+import co.fineants.member.application.MemberNotificationService;
 import co.fineants.member.domain.Member;
+import co.fineants.member.presentation.MemberNotificationRestController;
 import co.fineants.member.presentation.dto.response.MemberNotification;
 import co.fineants.member.presentation.dto.response.MemberNotificationResponse;
