+++ src/test/java/co/fineants/api/docs/member/MemberRestControllerDocsTest.java
@@ -24,11 +24,11 @@ import org.springframework.restdocs.payload.JsonFieldType;
 
 import co.fineants.TestDataFactory;
 import co.fineants.api.docs.RestDocsSupport;
-import co.fineants.api.domain.member.controller.MemberRestController;
-import co.fineants.member.application.MemberService;
 import co.fineants.api.global.util.ObjectMapperUtil;
+import co.fineants.member.application.MemberService;
 import co.fineants.member.domain.Member;
 import co.fineants.member.domain.Nickname;
+import co.fineants.member.presentation.MemberRestController;
 import co.fineants.member.presentation.dto.request.ProfileChangeServiceRequest;
 import co.fineants.member.presentation.dto.response.OauthMemberResponse;
 import co.fineants.member.presentation.dto.response.ProfileChangeResponse;
