+++ src/test/java/co/fineants/api/docs/member/SignUpRestControllerDocsTest.java
@@ -23,12 +23,12 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
 import co.fineants.api.docs.RestDocsSupport;
-import co.fineants.api.domain.member.controller.SignUpRestController;
+import co.fineants.api.global.util.ObjectMapperUtil;
+import co.fineants.member.application.MemberProfileFactory;
 import co.fineants.member.application.SignupService;
 import co.fineants.member.application.SignupValidatorService;
 import co.fineants.member.application.SignupVerificationService;
-import co.fineants.api.global.util.ObjectMapperUtil;
-import co.fineants.member.application.MemberProfileFactory;
+import co.fineants.member.presentation.SignUpRestController;
 
 class SignUpRestControllerDocsTest extends RestDocsSupport {
