+++ src/test/java/co/fineants/member/presentation/MemberNotificationRestControllerTest.java
@@ -1,4 +1,4 @@
-package co.fineants.api.domain.member.controller;
+package co.fineants.member.presentation;
 
 import static co.fineants.api.domain.notification.domain.entity.type.NotificationType.*;
 import static org.hamcrest.Matchers.*;
@@ -18,10 +18,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 
 import co.fineants.api.domain.common.money.Money;
-import co.fineants.member.application.MemberNotificationPreferenceService;
-import co.fineants.member.application.MemberNotificationService;
 import co.fineants.api.domain.notification.domain.entity.NotificationBody;
 import co.fineants.api.global.util.ObjectMapperUtil;
+import co.fineants.member.application.MemberNotificationPreferenceService;
+import co.fineants.member.application.MemberNotificationService;
 import co.fineants.member.presentation.dto.response.MemberNotification;
 import co.fineants.member.presentation.dto.response.MemberNotificationResponse;
 import co.fineants.support.controller.ControllerTestSupport;
